+++ angular.json
@@ -2,13 +2,14 @@
   "$schema": "./node_modules/@angular/cli/lib/config/schema.json",
   "version": 1,
   "newProjectRoot": "projects",
+  "defaultProject": "entity-state",
   "cli": {
     "packageManager": "yarn"
   },
   "projects": {
-    "entity-state-integration": {
-      "root": "",
-      "sourceRoot": "src",
+    "integration": {
+      "root": "integration",
+      "sourceRoot": ".",
       "projectType": "application",
       "prefix": "app",
       "schematics": {
@@ -20,17 +21,17 @@
         "build": {
           "builder": "@angular-devkit/build-angular:browser",
           "options": {
-            "outputPath": "dist/entity-state-integration",
-            "index": "src/index.html",
-            "main": "src/main.ts",
-            "polyfills": "src/polyfills.ts",
-            "tsConfig": "src/tsconfig.app.json",
+            "outputPath": "dist-integration/",
+            "index": "integration/index.html",
+            "main": "integration/main.ts",
+            "polyfills": "integration/polyfills.ts",
+            "tsConfig": "integration/tsconfig.app.json",
             "assets": [
-              "src/favicon.ico",
-              "src/assets"
+              "integration/favicon.ico",
+              "integration/assets"
             ],
             "styles": [
-              "src/styles.scss"
+              "integration/styles.scss"
             ],
             "scripts": []
           },
@@ -38,8 +39,8 @@
             "production": {
               "fileReplacements": [
                 {
-                  "replace": "src/environments/environment.ts",
-                  "with": "src/environments/environment.prod.ts"
+                  "replace": "integration/environments/environment.ts",
+                  "with": "integration/environments/environment.prod.ts"
                 }
               ],
               "optimization": true,
@@ -64,34 +65,34 @@
         "serve": {
           "builder": "@angular-devkit/build-angular:dev-server",
           "options": {
-            "browserTarget": "entity-state-integration:build"
+            "browserTarget": "integration:build"
           },
           "configurations": {
             "production": {
-              "browserTarget": "entity-state-integration:build:production"
+              "browserTarget": "integration:build:production"
             }
           }
         },
         "extract-i18n": {
           "builder": "@angular-devkit/build-angular:extract-i18n",
           "options": {
-            "browserTarget": "entity-state-integration:build"
+            "browserTarget": "integration:build"
           }
         },
         "test": {
           "builder": "@angular-devkit/build-angular:karma",
           "options": {
-            "main": "src/test.ts",
-            "polyfills": "src/polyfills.ts",
-            "tsConfig": "src/tsconfig.spec.json",
-            "karmaConfig": "src/karma.conf.js",
+            "main": "integration/test.ts",
+            "polyfills": "integration/polyfills.ts",
+            "tsConfig": "integration/tsconfig.spec.json",
+            "karmaConfig": "karma.conf.js",
             "styles": [
-              "src/styles.scss"
+              "integration/styles.scss"
             ],
             "scripts": [],
             "assets": [
-              "src/favicon.ico",
-              "src/assets"
+              "integration/favicon.ico",
+              "integration/assets"
             ]
           }
         },
@@ -99,37 +100,9 @@
           "builder": "@angular-devkit/build-angular:tslint",
           "options": {
             "tsConfig": [
-              "src/tsconfig.app.json",
-              "src/tsconfig.spec.json"
+              "integration/tsconfig.app.json",
+              "integration/tsconfig.spec.json"
             ],
-            "exclude": [
-              "**/node_modules/**"
-            ]
-          }
-        }
-      }
-    },
-    "entity-state-integration-e2e": {
-      "root": "e2e/",
-      "projectType": "application",
-      "prefix": "",
-      "architect": {
-        "e2e": {
-          "builder": "@angular-devkit/build-angular:protractor",
-          "options": {
-            "protractorConfig": "e2e/protractor.conf.js",
-            "devServerTarget": "entity-state-integration:serve"
-          },
-          "configurations": {
-            "production": {
-              "devServerTarget": "entity-state-integration:serve:production"
-            }
-          }
-        },
-        "lint": {
-          "builder": "@angular-devkit/build-angular:tslint",
-          "options": {
-            "tsConfig": "e2e/tsconfig.e2e.json",
             "exclude": [
               "**/node_modules/**"
             ]
@@ -138,32 +111,32 @@
       }
     },
     "entity-state": {
-      "root": "projects/entity-state",
-      "sourceRoot": "projects/entity-state/src",
+      "root": "",
+      "sourceRoot": "src",
       "projectType": "library",
-      "prefix": "lib",
+      "prefix": "entity",
       "architect": {
         "build": {
           "builder": "@angular-devkit/build-ng-packagr:build",
           "options": {
-            "tsConfig": "projects/entity-state/tsconfig.lib.json",
-            "project": "projects/entity-state/ng-package.json"
+            "tsConfig": "src/tsconfig.lib.json",
+            "project": "src/ng-package.json"
           }
         },
         "test": {
           "builder": "@angular-devkit/build-angular:karma",
           "options": {
-            "main": "projects/entity-state/src/test.ts",
-            "tsConfig": "projects/entity-state/tsconfig.spec.json",
-            "karmaConfig": "projects/entity-state/karma.conf.js"
+            "main": "src/test.ts",
+            "karmaConfig": "karma.conf.js",
+            "tsConfig": "src/tsconfig.spec.json"
           }
         },
         "lint": {
           "builder": "@angular-devkit/build-angular:tslint",
           "options": {
             "tsConfig": [
-              "projects/entity-state/tsconfig.lib.json",
-              "projects/entity-state/tsconfig.spec.json"
+              "src/tsconfig.lib.json",
+              "src/tsconfig.spec.json"
             ],
             "exclude": [
               "**/node_modules/**"
@@ -172,6 +145,5 @@
         }
       }
     }
-  },
-  "defaultProject": "entity-state-integration"
+  }
 }
