+++ integration/favicon.ico
