+++ integration/tslint.json
@@ -1,16 +1,16 @@
 {
-  "extends": "../../tslint.json",
+  "extends": "../tslint.json",
   "rules": {
     "directive-selector": [
       true,
       "attribute",
-      "lib",
+      "app",
       "camelCase"
     ],
     "component-selector": [
       true,
       "element",
-      "lib",
+      "app",
       "kebab-case"
     ]
   }
