--- projects/entity-state/ng-package.json
@@ -1,7 +1,0 @@
-{
-  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/entity-state",
-  "lib": {
-    "entryFile": "src/public_api.ts"
-  }
-}
