+++ src/lib/entity-state.ts
