+++ src/lib/internal.ts
