+++ src/tsconfig.lib.json
@@ -1,7 +1,7 @@
 {
-  "extends": "../../tsconfig.json",
+  "extends": "../tsconfig.json",
   "compilerOptions": {
-    "outDir": "../../out-tsc/lib",
+    "outDir": "../out-tsc/lib",
     "target": "es2015",
     "module": "es2015",
     "moduleResolution": "node",
